finish splitting hdk functions to own files
