+++ hdk-rust/src/api/entry_address.rs
@@ -1,0 +1,47 @@
+use holochain_core_types::entry::Entry;
+use error::ZomeApiResult;
+use holochain_core_types::cas::content::Address;
+use super::Dispatch;
+/// Reconstructs an address of the given entry data.
+/// This is the same value that would be returned if `entry_type_name` and `entry_value` were passed
+/// to the [commit_entry](fn.commit_entry.html) function and by which it would be retrievable from the DHT using [get_entry](fn.get_entry.html).
+/// This is often used to reconstruct an address of a `base` argument when calling [get_links](fn.get_links.html).
+/// # Examples
+/// ```rust
+/// # #![feature(try_from)]
+/// # extern crate hdk;
+/// # extern crate serde_json;
+/// # #[macro_use]
+/// # extern crate serde_derive;
+/// # extern crate holochain_core_types;
+/// # #[macro_use]
+/// # extern crate holochain_core_types_derive;
+/// # use hdk::error::ZomeApiResult;
+/// # use holochain_core_types::json::JsonString;
+/// # use holochain_core_types::error::HolochainError;
+/// # use holochain_core_types::entry::entry_type::AppEntryType;
+/// # use holochain_core_types::entry::AppEntryValue;
+/// # use holochain_core_types::entry::Entry;
+/// # use holochain_core_types::cas::content::Address;
+/// # fn main() {
+///
+/// #[derive(Serialize, Deserialize, Debug, DefaultJson)]
+/// pub struct Post {
+///     content: String,
+///     date_created: String,
+/// }
+///
+/// pub fn handle_post_address(content: String) -> ZomeApiResult<Address> {
+///     let post_entry = Entry::App("post".into(), Post {
+///         content,
+///         date_created: "now".into(),
+///     }.into());
+///
+///     hdk::entry_address(&post_entry)
+/// }
+///
+/// # }
+/// ```
+pub fn entry_address(entry: &Entry) -> ZomeApiResult<Address> {
+    Dispatch::EntryAddress.with_input(entry)
+}
