+++ hdk-rust/src/api/property.rs
@@ -1,0 +1,11 @@
+use error::{ZomeApiResult, ZomeApiError};
+
+/// NOT YET AVAILABLE
+// Returns a DNA property, which are defined by the DNA developer.
+// They are custom values that are defined in the DNA file
+// that can be used in the zome code for defining configurable behaviors.
+// (e.g. Name, Language, Description, Author, etc.).
+pub fn property<S: Into<String>>(_name: S) -> ZomeApiResult<String> {
+    Err(ZomeApiError::FunctionNotImplemented)
+}
+
